src/requirements.txt
black 
pylint
numpy  
scipy
notebook
jupyter_contrib_nbextensions 
ipykernel
autopep8
matplotlib
pandas
ipdb
pyyaml
networkx
seaborn
statsmodels